+++ cache.h
@@ -109,10 +109,10 @@ private:
     Tag_Array* m_TagArray;
     bool m_DebugMode;
     bool m_pLRU[NofIndex][CacheAssc-1];
-    unsigned int m_CacheRead;
-    unsigned int m_CacheWrite;
-    unsigned int m_CacheHit;
-    unsigned int m_CacheMiss;
+    double m_CacheRead;
+    double m_CacheWrite;
+    double m_CacheHit;
+    double m_CacheMiss;
 };
